@@ -1,11 +1,17 @@
 <?php
+namespace SentryExtra\v3;
 
+use MODX\Revolution\modX;
 use function Sentry\captureException;
 
 class SentryExtra
 {
-    public $modx = null;
+    /** @var modX $modx */
+    public $modx;
+
     public $namespace = 'sentryextra';
+
+    /** @var array $options */
     public $options = [];
 
     public function __construct(modX &$modx, array $options = [])
@@ -14,35 +20,32 @@ class SentryExtra
 
         $corePath = $this->getOption('core_path', $options, $this->modx->getOption('core_path', null, MODX_CORE_PATH) . 'components/sentryextra/');
 
-        /* loads some default paths for easier management */
-        $this->options = array_merge([
-            'namespace' => $this->namespace,
-            'corePath' => $corePath,
-            'modelPath' => $corePath . 'model/',
-            'snippetsPath' => $corePath . 'elements/snippets/',
-            'pluginsPath' => $corePath . 'elements/plugins/',
-        ], $options);
-
-        $this->modx->addPackage('sentryextra', $this->getOption('modelPath'));
+        $this->options = array_merge(
+            [
+                'namespace' => $this->namespace,
+                'corePath'  => $corePath,
+                'srcPath'   => $corePath . 'src/',
+                'modelPath' => $corePath . 'src/Model/',
+                'snippetsPath' => $corePath . 'elements/snippets/',
+                'pluginsPath' => $corePath . 'elements/plugins/',
+            ],
+            $options
+        );
         $this->modx->lexicon->load('sentryextra:default');
-        $this->autoload();
     }
 
-    protected function autoload()
-    {
-        require_once $this->getOption('corePath') . 'vendor/autoload.php';
-    }
 
     /**
      * Get a local configuration option or a namespaced system setting by key.
      *
-     * @param string $key The option key to search for.
-     * @param array $options An array of options that override local options.
-     * @param mixed $default The default value returned if the option is not found locally or as a
+     * @param  string  $key  The option key to search for.
+     * @param  array  $options  An array of options that override local options.
+     * @param  mixed  $default  The default value returned if the option is not found locally or as a
      * namespaced system setting; by default this value is null.
+     *
      * @return mixed The option value or the default value specified.
      */
-    public function getOption($key, $options = array(), $default = null)
+    public function getOption(string $key, $options = [], $default = null)
     {
         $option = $default;
         if (!empty($key) && is_string($key)) {
@@ -56,7 +59,7 @@ class SentryExtra
         }
         return $option;
     }
-
+    
     public function log($level, $msg, $file = '', $line = ''): void
     {
         captureException(new \ErrorException($msg, 0, $level, $file, $line));
