@@ -1,12 +1,16 @@
 <?php
 
-namespace SentryExtra\Event;
+namespace SentryExtra\v3\Event;
+
+use SentryExtra\v3\ErrorHandler;
+use function Sentry\init;
 
 class OnMODXInit extends Event
 {
     public function run()
     {
         $dsn = $this->getOption('sentryextra.dsn');
+        $environment = $this->getOption('sentryextra.environment', 'production');
         $traces_sample_rate = (float) $this->getOption('sentryextra.traces_sample_rate', null, 0.1);
         if ($traces_sample_rate > 1) {
             $traces_sample_rate = 1;
@@ -15,11 +19,12 @@ class OnMODXInit extends Event
             $traces_sample_rate = 0;
         }
         if ($dsn) {
-            \Sentry\init([
+            init([
                 'dsn' => $dsn,
                 'traces_sample_rate' => $traces_sample_rate,
+                'environment' => $environment,
             ]);
-            $this->modx->errorHandler = new \SentryExtra\ErrorHandler($this->modx, []);
+            $this->modx->errorHandler = new ErrorHandler($this->modx, []);
             set_error_handler(
                 array($this->modx->errorHandler, 'handleError'),
                 $this->getOption('error_handler_types', error_reporting(), true)
