@@ -1,15 +1,15 @@
 <?php
 
-namespace SentryExtra\Event;
+namespace SentryExtra\v3\Event;
 
-class OnLoadWebPageCache extends Event
+class OnWebPageInit extends Event
 {
     public function run()
     {
         $dsn = $this->getOption('sentryextra.dsn');
         if ($dsn) {
             if (method_exists($this->modx->parser, 'endTransaction')) {
-                $this->modx->parser->endTransaction();
+                $this->modx->parser->startTransaction();
             }
         }
     }
